runtime clean imports, disable row-wise min/max
