+++ runtime-rust/src/components/kth_raw_sample_moment.rs
@@ -4,14 +4,17 @@ use crate::base::NodeArguments;
 use whitenoise_validator::base::{Value, get_argument, ArrayND};
 use crate::components::Evaluable;
 use whitenoise_validator::proto;
-use ndarray::ArrayD;
+use ndarray::{ArrayD, Array};
 use crate::utilities::utilities::get_num_columns;
+use crate::components::mean::mean;
+
+use std::convert::TryFrom;
 
 impl Evaluable for proto::KthRawSampleMoment {
     fn evaluate(&self, arguments: &NodeArguments) -> Result<Value> {
         let data = get_argument(&arguments, "data")?.get_arraynd()?.get_f64()?;
         let k = get_argument(&arguments, "k")?.get_first_i64()?;
-        Ok(Value::ArrayND(ArrayND::F64(kth_raw_sample_moment(&data, &k))))
+        Ok(kth_raw_sample_moment(data, &k)?.into())
     }
 }
 
@@ -27,7 +30,7 @@ impl Evaluable for proto::KthRawSampleMoment {
 ///
 /// # Example
 /// ```
-/// use ndarray::{ArrayD, arr2};
+/// use ndarray::{ArrayD, arr2, arr1};
 /// use whitenoise_runtime::components::kth_raw_sample_moment::kth_raw_sample_moment;
 /// let data: ArrayD<f64> = arr2(&[ [1., 1., 1.], [2., 4., 6.] ]).into_dyn();
 /// let second_moments = kth_raw_sample_moment(&data, &2).unwrap();
@@ -38,7 +41,9 @@ pub fn kth_raw_sample_moment(data: &ArrayD<f64>, k: &i64) -> Result<ArrayD<f64>>
     let mut data = data.clone();
 
     let num_columns = get_num_columns(&data)?;
-
+    let k = match i32::try_from(*k) {
+        Ok(v) => v, Err(_) => return Err("k: invalid size".into())
+    };
     // iterate over the generalized columns
     data.gencolumns_mut().into_iter()
         // for each pairing, iterate over the cells
@@ -46,8 +51,5 @@ pub fn kth_raw_sample_moment(data: &ArrayD<f64>, k: &i64) -> Result<ArrayD<f64>>
             // mutate the cell via the operator
             .for_each(|v| *v = v.powi(k)));
 
-    Ok(data)
-    let data_vec: Vec<f64> = data.clone().into_dimensionality::<Ix1>().unwrap().to_vec();
-    let data_to_kth_power: Vec<f64> = data_vec.iter().map(|x| x.powf(*k as f64)).collect();
-    return mean(&arr1(&data_to_kth_power).into_dyn());
+    mean(&data)
 }
