+++ runtime-rust/src/components/mean.rs
@@ -1,7 +1,7 @@
 use whitenoise_validator::errors::*;
 
 use crate::base::NodeArguments;
-use whitenoise_validator::base::{Value, ArrayND, get_argument};
+use whitenoise_validator::base::{Value, get_argument};
 use crate::components::Evaluable;
 use ndarray::{ArrayD, Array};
 use crate::utilities::utilities::get_num_columns;
@@ -9,7 +9,7 @@ use whitenoise_validator::proto;
 
 impl Evaluable for proto::Mean {
     fn evaluate(&self, arguments: &NodeArguments) -> Result<Value> {
-        Ok(mean(&get_argument(&arguments, "data")?.get_arraynd()?.get_f64()?.clone())?.into())
+        Ok(mean(get_argument(&arguments, "data")?.get_arraynd()?.get_f64()?)?.into())
     }
 }
