+++ runtime-rust/src/components/variance.rs
@@ -1,7 +1,7 @@
 use whitenoise_validator::errors::*;
 
 use crate::base::NodeArguments;
-use whitenoise_validator::base::{Value, ArrayND, get_argument};
+use whitenoise_validator::base::{Value, get_argument};
 use crate::components::Evaluable;
 use ndarray::{ArrayD, Array};
 use crate::utilities::utilities::get_num_columns;
